Addition d'un mapper pour faire la convertion entre DomainModel et ViewModel;
Changement du nom de la base de données et de la collection;

Quelques champs ont été rénomés.

+++ AppCV-Fiches/DAL_CV-Fiches/Models/Graph/Technologie.cs
@@ -16,6 +16,6 @@ namespace DAL_CV_Fiches.Models.Graph
         public CategorieDeTechnologie Categorie { get; set; }
 
         [EdgeProperty]
-        public int MoisDExperience { get; set; }
+        public double MoisDExperience { get; set; }
     }
 }
